Added handling for the SHIFT, ALT and CTRL keys.
The camera movements that rely on them are still pretty kludgy, but at
least you can do them now. I also added handling for shift+mousewheel,
which now zooms the camera (previously this was just the mousewheel, no
shift key) and the plain mousewheel now dollies the camera in or out by
a fixed amount.