Added methods to help with texture uploading.
You can now downsample an image in place, prior to uploading it as a texture.

You can specify a target type for the texture upload (mainly useful when uploading a matte image, so you can specify a type of GL_ALPHA, for example).

Added the takePixels method, which returns the memory for the block of pixels and transfers ownership of the memory to the receiver as well.