Made all Vec3 functions templatized, because I realised I could.
The only thing stopping me before was the need to call type-specific methods like powf, sqrtf, and so on. But I realised that <cmath> declares template versions of these that invoke the right version for the type you pass in. So if we use those, there's nothing to stop us making templates out of all the remaining functions (which is nice, because it means I don't have to write new versions of them if I want to use doubles instead of floats; or if I one day want to use long doubles...).

Next step will be to do the same to Vec2, Vec4 and Ray3.