Some arcball corrections and enhancements.
The z value of 0 that I was passing in to the unproject function was incorrect. 0 means on the near clipping plane, but what I want is 0.5 (halfway between the near and far clipping planes).

Added an arcballMove function which attempts to move the camera in the plane perpendicular to the view direction, rather than in the global X-Y plane. It doesn't work correctly - see the WTF message in the code - and I'm too tired to figure out why right now. Hopefully it'll make more sense after some rest.