Cleaned up the arcball code.
Roll works well enough now, but could be improved further still. In particular, when you reach the edge of the "ball" it just stops moving currently. It should transition to a virtual flat surface at that point instead and use that to determine how much to roll by.

Move works... ish. The translation seems to happen before the camera transform is applied, so it translates along the world axes rather than the camera-local ones.