Made the Camera class into a pure virtual class and moved it's functionality into a subclass called BaseCamera.
Updated all the examples which define their own camera classes to inherit from BaseCamera instead. Added an include line for the BaseCamera header into vgl.h. Added it to the Makefile. Tested. :-)

This is another precursor to integrating the arcball camera (and the ortho camera from the imageview example too, now that you mention it) into the main VGL codebase.