Refactored the arcball camera to override the roll and move methods from BaseCamera, now that their signature matches what we need.
Had to make some slight modifications to the arcball camera methods so that they work correctly with the values the default viewer gives it. The ArcballViewer class had been inverting the y coordinates before invoking the camera methods, but that should really have been done by the camera itself. Well, now it *is*.

Was able to get rid of the ArcballViewer class altogether after that, which simplifies the example code nicely.