Added an unproject method in a new file, vgl_funcs.h.
vgl_funcs.h is for miscellanous helper functions which use classes from VGL. vgl_utils.h remains solely for functions which don't depend on other parts of VGL (because it's used by them - it's one of the roots of the dependency tree).

I added a comment to vgl_utils.h explaining this.