Modified the other examples to use the default camera instead of providing their own.
Where appropriate, of course.

The arcball example still explicitly provides a camera, even though it's actually the same as the default camera we'd get. But the point of that example is to show how to create and use an ArcballCamera, so it seemed counter-productive to take it out...

Naturally the imageview example still uses an OrthoCamera, too.